Coconut (#23)
## About The Pull Request

Adds palm tree and coconut as a hydroponics feature.
This is going in as a PR because I'm not sure how watertight this is, I
don't play botany enough. Normally test merges are for things that NEED
multiple players to test them, but, ehhhh

Grow coconuts, throw them at things, 50% chance they'll break.
Inexplicably you can crack a coconut using it in hand too (probably need
a better tool action for that, maybe a hatchet?)
Cracked coconuts are food, can be juiced, and put them into a food
processor for dessicated coconut (currently not used by anything)

## Why It's Good For The Game


https://github.com/user-attachments/assets/7d2be78d-e92b-4183-8107-ff97a6f84ed8

## Changelog

:cl:
add: Added coconut, thanks to Guffy for sprites
/:cl: