Relic Artifacts initial commit (#32)
## About The Pull Request
This pull request is about making relics more interesting by
implementing an evolving node graph for effects and reactions. Giving
science an endless goal worth pursuing, and giving plenty of reasons to
collaborate (willingly or otherwise) with other departments.
## Why It's Good For The Game
Relics right now are an afterthought in the science department, along
with 2 of the other jobs (cytology, virology). The science department
needs something that meaningfully adds to the dynamic of the station.
This overarching tweak to relics allows them to be agents of chaos for
the science team to manage, and gives a clear generalist secondary
objective to pursue. The place of relic experimentation is entirely on
the department's hubris, meaning the system is engaged with on a
player's own terms and as much as they'd like.
## Changelog
:cl: ArchytasBird
add: relics now have a procedurally generated web of effects that can
happen when introduced to various stimuli
add: relic cargo order to order more relics once the station is picked
clean
add: relic pinpointer to monitor relic states & assist in finding
fruitful stimuli
add: relic pinpointer to science wardrobe, crafting and autolathe
recepies
image: plushie_archy
code: small chance for any relic to have the plushie_archy sprite
applied
refactor: minor extension to how relics handle variants
admin: added some admin alerts for potentially disastrous relic effects

/:cl: